# openapi/v1/api.yaml
openapi: 3.0.0
paths:
...
  /v1/auth/login:
    post:
      operationId: AuthController_login
      summary: Autenticar usuário e obter tokens de acesso
      parameters: []
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairDto'
        '401':
          description: Credenciais inválidas
        '429':
          description: Muitas tentativas — tente mais tarde
      tags:
        - auth
  /v1/auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: Renovar tokens usando refresh token
      parameters: []
      responses:
        '200':
          description: Tokens renovados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairDto'
        '401':
          description: Refresh token inválido ou expirado
      tags:
        - auth
  /v1/auth/logout:
    post:
      operationId: AuthController_logout
      summary: Encerrar sessão do usuário
      parameters: []
      responses:
        '204':
          description: Logout realizado com sucesso
        '401':
          description: Não autenticado
      tags:
        - auth
      security:
        - bearer: []
...
  /v1/drivers:
    get:
      operationId: DriversController_findAll
      summary: List all drivers with cursor-based pagination
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
    post:
      operationId: DriversController_create
      summary: Create a new driver
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
  /v1/drivers/{id}:
    get:
      operationId: DriversController_findById
      summary: Get driver by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
    patch:
      operationId: DriversController_update
      summary: Update driver
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
    delete:
      operationId: DriversController_delete
      summary: Soft delete a driver
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
  /v1/drivers/{id}/availability:
    get:
      operationId: DriversController_checkAvailability
      summary: Check driver availability
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
  /v1/drivers/{id}/history:
    get:
      operationId: DriversController_getHistory
      summary: Get driver trip history (last 20)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - drivers
      security:
        - bearer: []
...
  /v1/fuel-records:
    get:
      operationId: FuelController_findAll
      summary: List all fuel records with cursor-based pagination
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - fuel
      security:
        - bearer: []
    post:
      operationId: FuelController_create
      summary: Create a fuel record — validates odometer, detects anomaly
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - fuel
      security:
        - bearer: []
  /v1/fuel-records/anomalies:
    get:
      operationId: FuelController_findAnomalies
      summary: List fuel records with anomaly flag
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - fuel
      security:
        - bearer: []
  /v1/fuel-records/stats:
    get:
      operationId: FuelController_getStats
      summary: Get fuel consumption statistics
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - fuel
      security:
        - bearer: []
  /v1/fuel-records/{id}:
    get:
      operationId: FuelController_findById
      summary: Get fuel record by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - fuel
      security:
        - bearer: []
    patch:
      operationId: FuelController_update
      summary: Update non-critical fuel record fields
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - fuel
      security:
        - bearer: []
...
  /v1/finance/summary:
    get:
      operationId: FinanceController_getSummary
      summary: Get financial summary for a period
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - finance
      security:
        - bearer: []
  /v1/finance/revenue-by-client:
    get:
      operationId: FinanceController_getRevenueByClient
      summary: Get revenue breakdown by client
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - finance
      security:
        - bearer: []
  /v1/finance/cost-by-vehicle:
    get:
      operationId: FinanceController_getCostByVehicle
      summary: Get cost breakdown by vehicle
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - finance
      security:
        - bearer: []
...
  /v1/audit:
    get:
      operationId: AuditController_findAll
      summary: List audit logs for the current tenant (CEO/Manager only)
      description: Cursor-based pagination ordered by createdAt DESC. Strictly tenant-isolated — cannot access other tenant data.
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - audit
      security:
        - bearer: []
  /v1/audit/{id}:
    get:
      operationId: AuditController_findById
      summary: Get audit log entry by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - audit
      security:
        - bearer: []
...
info:
  title: Gestão Fretamento Pro API
  description: API REST para gestão de fretamento empresarial
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    TokenUserDto:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        email:
          type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - tenantId
        - name
        - email
        - permissions
    TokenPairDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiresAt:
          format: date-time
          type: string
        refreshTokenExpiresAt:
          format: date-time
          type: string
        user:
          $ref: '#/components/schemas/TokenUserDto'
      required:
        - accessToken
        - refreshToken
        - accessTokenExpiresAt
        - refreshTokenExpiresAt
        - user
...